Conflicts:
	sp/volna2hdf5/Makefile
	sp/volna2hdf5/volna2hdf5.cpp

+++ sp/initV_formula.h
@@ -1,4 +1,4 @@
-inline void initV_formula(float *coords, float *values, const float *time) {
+inline void initV_formula(float *coords, float *values, const double *time) {
 //UNCOMMENT THE ONES USED
   // float x = coords[0];
   // float y = coords[1];
